requirements.txt
Pillow==9.1.0
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.24.1
torch==2.0.1+cu118
torchvision==0.15.2+cu118
